updated baths to decimal

+++ backend/db/migrations/20220630230354-create-room.js
@@ -70,11 +70,11 @@ module.exports = {
         allowNull: false,
       },
       beds: {
-        type: Sequelize.DECIMAL(3,1),
+        type: Sequelize.INTEGER,
         allowNull: false,
       },
       baths: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(3,1),
         allowNull: false,
       },
       createdAt: {
